+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- hnefatafl-client: add win % to the player statistics.
+
 ## [4.4.1] - 2025-12-05
 
 ### Fixed